cluster-ui: add explain plan tab for statement insights
Closes #88093

This commit adds a new tab to the statement insight details
page. The page is now separated into two tabs:
1. Overview - what was already  on the page (insights details)
2. Explain Plan - the explain plan for the stmt being viewed

Release note (ui change): new tab view added to the stmt insight
details page:
1. Overview - what was already  on the page (insights details)
2. Explain Plan - the explain plan for the stmt being viewed
