+++ pkg/ui/workspaces/cluster-ui/src/api/insightsApi.ts
@@ -591,6 +591,7 @@ type ExecutionInsightsResponseRow = {
   causes: string[];
   problem: string;
   index_recommendations: string[];
+  plan_gist: string;
 };
 
 export type StatementInsights = StatementInsightEvent[];
@@ -630,6 +631,7 @@ function getStatementInsightsFromClusterExecutionInsightsResponse(
       problem: row.problem,
       indexRecommendations: row.index_recommendations,
       insights: null,
+      planGist: row.plan_gist,
     };
   });
 }
@@ -669,6 +671,7 @@ const statementInsightsQuery: InsightQuery<
       index_recommendations,
       problem,
       causes,
+      plan_gist,
       row_number()                          OVER (
         PARTITION BY txn_fingerprint_id
         ORDER BY end_time DESC
