+++ pkg/ui/workspaces/cluster-ui/src/insights/workloadInsightDetails/insightsDetails.module.scss
@@ -1,0 +1,3 @@
+.section {
+  padding: 12px 24px 12px 0px;
+}
