+++ pkg/ui/workspaces/cluster-ui/src/insights/workloadInsightDetails/statementInsightDetails.tsx
@@ -7,120 +7,39 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0, included in the file
 // licenses/APL.txt.
-import React, { useContext, useMemo } from "react";
+import React, { useState } from "react";
 import Helmet from "react-helmet";
 import { RouteComponentProps } from "react-router-dom";
 import { ArrowLeft } from "@cockroachlabs/icons";
-import { Col, Row } from "antd";
+import { Row, Col, Tabs } from "antd";
+import "antd/lib/tabs/style";
 import "antd/lib/col/style";
 import "antd/lib/row/style";
 import { Button } from "src/button";
 import { Loading } from "src/loading";
 import { SqlBox, SqlBoxSize } from "src/sql";
-import { SummaryCard, SummaryCardItem } from "src/summaryCard";
-import { capitalize, Duration } from "src/util";
-import { DATE_FORMAT_24_UTC } from "src/util/format";
 import { getMatchParamByName } from "src/util/query";
-import {
-  InsightsSortedTable,
-  makeInsightsColumns,
-} from "src/insightsTable/insightsTable";
-import { populateStatementInsightsFromProblemAndCauses } from "../utils";
-import {
-  InsightNameEnum,
-  StatementInsightEvent,
-  executionDetails,
-  InsightRecommendation,
-} from "../types";
+import { StatementInsightEvent } from "../types";
 import { InsightsError } from "../insightsErrorComponent";
-
 import classNames from "classnames/bind";
+
 import { commonStyles } from "src/common";
-import insightTableStyles from "src/insightsTable/insightsTable.module.scss";
-import summaryCardStyles from "src/summaryCard/summaryCard.module.scss";
-import { CockroachCloudContext } from "../../contexts";
+import { getExplainPlanFromGist } from "src/api/decodePlanGistApi";
+import { StatementInsightDetailsOverviewTab } from "./statementInsightDetailsOverviewTab";
 
-const tableCx = classNames.bind(insightTableStyles);
-const summaryCardStylesCx = classNames.bind(summaryCardStyles);
+// Styles
+import insightsDetailsStyles from "src/insights/workloadInsightDetails/insightsDetails.module.scss";
 
-const insightsTableData = (
-  insightDetails: StatementInsightEvent | null,
-): InsightRecommendation[] => {
-  if (!insightDetails) return [];
+const cx = classNames.bind(insightsDetailsStyles);
 
-  const recs: InsightRecommendation[] = [];
-  let rec: InsightRecommendation;
-  const execDetails: executionDetails = {
-    statement: insightDetails.query,
-    fingerprintID: insightDetails.statementFingerprintID,
-    retries: insightDetails.retries,
-  };
-  insightDetails.insights.forEach(insight => {
-    switch (insight.name) {
-      case InsightNameEnum.highContention:
-        rec = {
-          type: "HighContention",
-          execution: execDetails,
-          details: {
-            duration: insightDetails.elapsedTimeMillis,
-            description: insight.description,
-          },
-        };
-        break;
-      case InsightNameEnum.failedExecution:
-        rec = {
-          type: "FailedExecution",
-        };
-        break;
-      case InsightNameEnum.highRetryCount:
-        rec = {
-          type: "HighRetryCount",
-          execution: execDetails,
-          details: {
-            description: insight.description,
-          },
-        };
-        break;
-      case InsightNameEnum.planRegression:
-        rec = {
-          type: "PlanRegression",
-          execution: execDetails,
-          details: {
-            description: insight.description,
-          },
-        };
-        break;
-      case InsightNameEnum.suboptimalPlan:
-        rec = {
-          type: "SuboptimalPlan",
-          database: insightDetails.databaseName,
-          execution: {
-            ...execDetails,
-            indexRecommendations: insightDetails.indexRecommendations,
-          },
-          details: {
-            description: insight.description,
-          },
-        };
-        break;
-      default:
-        rec = {
-          type: "Unknown",
-          details: {
-            duration: insightDetails.elapsedTimeMillis,
-            description: insight.description,
-          },
-        };
-        break;
-    }
-    recs.push(rec);
-  });
-  return recs;
-};
-
+enum TabKeysEnum {
+  OVERVIEW = "overview",
+  EXPLAIN = "explain",
+}
 export interface StatementInsightDetailsStateProps {
   insightEventDetails: StatementInsightEvent;
   insightError: Error | null;
+  isTenant?: boolean;
 }
 
 export type StatementInsightDetailsProps = StatementInsightDetailsStateProps &
@@ -128,129 +47,89 @@ export type StatementInsightDetailsProps = StatementInsightDetailsStateProps &
 
 export const StatementInsightDetails: React.FC<
   StatementInsightDetailsProps
-> = ({ history, insightEventDetails, insightError, match }) => {
-  const isCockroachCloud = useContext(CockroachCloudContext);
+> = ({ history, insightEventDetails, insightError, match, isTenant }) => {
+  const [explain, setExplain] = useState<string>(null);
 
   const prevPage = (): void => history.goBack();
 
-  const insightsColumns = useMemo(
-    () => makeInsightsColumns(isCockroachCloud),
-    [isCockroachCloud],
-  );
+  const onTabClick = (key: TabKeysEnum) => {
+    if (
+      !isTenant &&
+      key === TabKeysEnum.EXPLAIN &&
+      insightEventDetails?.planGist &&
+      !explain
+    ) {
+      // Get the explain plan.
+      getExplainPlanFromGist({ planGist: insightEventDetails.planGist }).then(
+        res => {
+          setExplain(res.explainPlan || res.error);
+        },
+      );
+    }
+  };
 
-  const insightDetailsArr = useMemo(
-    () => populateStatementInsightsFromProblemAndCauses([insightEventDetails]),
-    [insightEventDetails],
-  );
-  const insightDetails = insightDetailsArr.length ? insightDetailsArr[0] : null;
-  const tableData = insightsTableData(insightDetails);
+  const executionID = getMatchParamByName(match, "id");
 
   return (
     <div>
       <Helmet title={"Details | Insight"} />
+      <Button
+        onClick={prevPage}
+        type="unstyled-link"
+        size="small"
+        icon={<ArrowLeft fontSize={"10px"} />}
+        iconPosition="left"
+        className={commonStyles("small-margin")}
+      >
+        Insights
+      </Button>
+      <h3 className={commonStyles("base-heading", "no-margin-bottom")}>
+        Statement Execution ID: {executionID}
+      </h3>
       <div>
-        <Button
-          onClick={prevPage}
-          type="unstyled-link"
-          size="small"
-          icon={<ArrowLeft fontSize={"10px"} />}
-          iconPosition="left"
-          className={commonStyles("small-margin")}
-        >
-          Insights
-        </Button>
-        <h3
-          className={commonStyles("base-heading", "no-margin-bottom")}
-        >{`Statement Execution ID: ${String(
-          getMatchParamByName(match, "id"),
-        )}`}</h3>
-      </div>
-      <section>
         <Loading
           loading={insightEventDetails == null}
           page={"Transaction Insight details"}
           error={insightError}
           renderError={() => InsightsError()}
         >
-          <section className={tableCx("section")}>
-            <Row gutter={24}>
-              <Col className="gutter-row" span={24}>
-                <SqlBox size={SqlBoxSize.custom} value={insightDetails.query} />
+          <section className={cx("section")}>
+            <Row>
+              <Col span={24}>
+                <SqlBox
+                  size={SqlBoxSize.custom}
+                  value={insightEventDetails?.query}
+                />
               </Col>
-            </Row>
-            <Row gutter={24}>
-              <Col className="gutter-row" span={12}>
-                <SummaryCard>
-                  <SummaryCardItem
-                    label="Start Time"
-                    value={insightDetails.startTime.format(DATE_FORMAT_24_UTC)}
-                  />
-                  <SummaryCardItem
-                    label="End Time"
-                    value={insightDetails.endTime.format(DATE_FORMAT_24_UTC)}
-                  />
-                  <SummaryCardItem
-                    label="Elapsed Time"
-                    value={Duration(insightDetails.elapsedTimeMillis * 1e6)}
-                  />
-                  <SummaryCardItem
-                    label="Rows Read"
-                    value={insightDetails.rowsRead}
-                  />
-                  <SummaryCardItem
-                    label="Rows Written"
-                    value={insightDetails.rowsWritten}
-                  />
-                  <SummaryCardItem
-                    label="Transaction Priority"
-                    value={capitalize(insightDetails.priority)}
-                  />
-                  <SummaryCardItem
-                    label="Full Scan"
-                    value={capitalize(String(insightDetails.isFullScan))}
-                  />
-                </SummaryCard>
-              </Col>
-              <Col className="gutter-row" span={12}>
-                <SummaryCard>
-                  <SummaryCardItem
-                    label="Transaction Retries"
-                    value={insightDetails.retries}
-                  />
-                  {insightDetails.lastRetryReason && (
-                    <SummaryCardItem
-                      label="Last Retry Reason"
-                      value={insightDetails.lastRetryReason.toString()}
-                    />
-                  )}
-                  <p
-                    className={summaryCardStylesCx("summary--card__divider")}
-                  />
-                  <SummaryCardItem
-                    label="Session ID"
-                    value={String(insightDetails.sessionID)}
-                  />
-                  <SummaryCardItem
-                    label="Transaction Fingerprint ID"
-                    value={String(insightDetails.transactionFingerprintID)}
-                  />
-                  <SummaryCardItem
-                    label="Transaction Execution ID"
-                    value={String(insightDetails.transactionID)}
-                  />
-                  <SummaryCardItem
-                    label="Statement Fingerprint ID"
-                    value={String(insightDetails.statementFingerprintID)}
-                  />
-                </SummaryCard>
-              </Col>
-            </Row>
-            <Row gutter={24} className={tableCx("margin-bottom")}>
-              <InsightsSortedTable columns={insightsColumns} data={tableData} />
             </Row>
           </section>
+          <Tabs
+            className={commonStyles("cockroach--tabs")}
+            defaultActiveKey={TabKeysEnum.OVERVIEW}
+            onTabClick={onTabClick}
+          >
+            <Tabs.TabPane tab="Overview" key={TabKeysEnum.OVERVIEW}>
+              <StatementInsightDetailsOverviewTab
+                insightEventDetails={insightEventDetails}
+              />
+            </Tabs.TabPane>
+            {!isTenant && (
+              <Tabs.TabPane tab="Explain Plan" key={TabKeysEnum.EXPLAIN}>
+                <section className={cx("section")}>
+                  <Row gutter={24}>
+                    <Col span={24}>
+                      <SqlBox
+                        value={explain || "Not available."}
+                        size={SqlBoxSize.custom}
+                      />
+                    </Col>
+                  </Row>
+                </section>
+              </Tabs.TabPane>
+            )}
+          </Tabs>
         </Loading>
-      </section>
+      </div>
     </div>
   );
 };
