+++ pkg/ui/workspaces/cluster-ui/src/insightsTable/insightsTable.module.scss
@@ -31,11 +31,6 @@
   margin-bottom: 30px;
 }
 
-.section {
-  flex: 0 0 auto;
-  padding: 12px 24px 12px 0px;
-}
-
 .inline {
   display: inline-flex;
 }
